fix: move extension to provider
